Cleaned up and added backend integration for deck creation, deletion UI. added partial backend integration for edit deck UI

+++ app/src/main/java/com/psu/capstonew17/pdxaslapp/CreateEditDeleteDeckActivity.java
@@ -2,50 +2,45 @@
 
 package com.psu.capstonew17.pdxaslapp;
 
-import java.util.Iterator;
 import java.util.List;
 import java.util.ListIterator;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
 import android.widget.RadioButton;
-import android.widget.LinearLayout;
 import android.widget.RadioGroup;
+import android.widget.Toast;
 
 import com.psu.capstonew17.backend.api.Deck;
-import com.psu.capstonew17.pdxaslapp.FrontEndTestStubs.TestingStubs;
-import com.psu.capstonew17.pdxaslapp.FrontEndTestStubs.testDeck;
+import com.psu.capstonew17.backend.api.DeckManager;
+import com.psu.capstonew17.backend.data.ExternalDeckManager;
 
 public class CreateEditDeleteDeckActivity extends BaseActivity{
-    private Button bttEdit, bttDlt, bttCrt;
     private RadioGroup deckRG;
     private List<Deck> decks;
+    private DeckManager deckManager;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_create_edit_delete_deck);
 
-        decks = TestingStubs.manyDecks();
-        ListIterator<Deck> deckIterator = decks.listIterator();
+        deckManager = ExternalDeckManager.getInstance(this);
         deckRG = (RadioGroup) findViewById(R.id.deckRButtons);
         populateRadioGroup(deckRG);
-
     }
 
     public void populateRadioGroup(RadioGroup deckRG){
+        deckRG.clearCheck();
         deckRG.removeAllViews();
+        decks = deckManager.getDecks(null);
         ListIterator<Deck> deckIterator = decks.listIterator();
 
-        while (deckIterator.hasNext()) {
-            Deck curr = deckIterator.next();
-
+        for (int i = 0; i < decks.size(); i++) {
             RadioButton currRad = new RadioButton(this);
-            currRad.setId(deckIterator.previousIndex());
-            currRad.setText(curr.getName());
+            currRad.setId(i);
+            currRad.setText(decks.get(i).getName());
             deckRG.addView(currRad);
         }
     }
@@ -59,20 +54,32 @@ public class CreateEditDeleteDeckActivity extends BaseActivity{
     //TODO: force the user to select a deck if they press this
     public void onDeleteClicked(View view) {
         int index = deckRG.getCheckedRadioButtonId();
-        Deck selectedDeck = decks.get(index);
-        decks.remove(index);
-        deckRG.clearCheck();
-        selectedDeck.delete();
-        selectedDeck.commit();
-        populateRadioGroup(deckRG);
-
+        if(index == -1) {
+            Toast.makeText(this, "ERROR Select a Deck", Toast.LENGTH_SHORT).show();
+        } else {
+            Deck selectedDeck = decks.get(index);
+            selectedDeck.delete();
+            selectedDeck.commit();
+            populateRadioGroup(deckRG);
+        }
     }
 
     //TODO: force the user to select a deck if they press this
     public void onEditClicked(View view) {
-        Intent intent;
-        intent = new Intent(this, EditDeckActivity.class);
-        intent.putExtra("checkedIndex", deckRG.getCheckedRadioButtonId());
-        startActivity(intent);
+        int index = deckRG.getCheckedRadioButtonId();
+        if(index == -1) {
+            Toast.makeText(this, "ERROR Select a Deck", Toast.LENGTH_SHORT).show();
+        } else {
+            Intent intent;
+            intent = new Intent(this, EditDeckActivity.class);
+            intent.putExtra("checkedIndex", index);
+            startActivity(intent);
+        }
+    }
+
+    @Override
+    protected void onResume(){
+        super.onResume();
+        populateRadioGroup(deckRG);
     }
 }
